Refactor check register modules for modularity
