+++ check_register/__init__.py
@@ -1,1 +1,4 @@
-from .parser import CheckEntry, CheckRegisterParser
+from .models import CheckEntry
+from .parser import CheckRegisterParser
+from .outputs import write_csv, write_json, write_payee_quadtree_html
+from .stats import sanity, month_rollups
